Additional fixes
Took care of the n/a issue with the motion derivatives, and added a check in postprep

+++ PostPrep.sh
@@ -144,40 +144,40 @@ for sub in ${SubID[@]}; do
 		for fun in ${FuncName[@]}; do
 			#
 			# Starting first with the preprocessed function: 1) convert to brik/head,
-			3dcopy ${PrepDir}/${sub}/func/${sub}_task-${fun}_space-${temp}_desc-preproc_bold.nii.gz ${PrepDir}/${sub}/afni/${sub}_${temp}_preproc_${fun}
-			# 2) calculate the voxel-level mean, 3) scale to the mean
-			3dTstat -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-preproc_${fun}+tlrc ${PrepDir}/${sub}/afni/${sub}_${temp}_preproc_${fun}+tlrc
-			# 3) scale to the mean
-			3dcalc -a ${PrepDir}/${sub}/afni/${sub}_${temp}_preproc_${fun}+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-preproc_${fun}+tlrc \
+			if [ ! -f ${PrepDir}/${sub}/afni/${sub}_${temp}_preproc_${fun}+tlrc.BRIK ]; then
+				3dcopy ${PrepDir}/${sub}/func/${sub}_task-${fun}_space-${temp}_desc-preproc_bold.nii.gz ${PrepDir}/${sub}/afni/${sub}_${temp}_preproc_${fun}
+				# 2) calculate the voxel-level mean, 3) scale to the mean
+				3dTstat -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-preproc_${fun}+tlrc ${PrepDir}/${sub}/afni/${sub}_${temp}_preproc_${fun}+tlrc
+				# 3) scale to the mean
+				3dcalc -a ${PrepDir}/${sub}/afni/${sub}_${temp}_preproc_${fun}+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-preproc_${fun}+tlrc \
 				-expr 'min(200, a/b*100)*step(a)*step(b)' \
 				-prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-preproc_${fun}+tlrc
-			# Clean up mean image
-			rm -f ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-preproc_${fun}+tlrc*
-			#
-			#
-			# Now copy over the brain mask into brik/head format
-			3dcopy ${PrepDir}/${sub}/func/${sub}_task-${fun}_space-${temp}_desc-brain_mask.nii.gz ${PrepDir}/${sub}/afni/${sub}_${temp}_brainmask_${fun}
-			#
-			#
-			# Finally copy over the AROMAe'd data and scale like before
-			3dcopy ${PrepDir}/${sub}/func/${sub}_task-${fun}_space-${temp}_desc-smoothAROMAnonaggr_bold.nii.gz ${PrepDir}/${sub}/afni/${sub}_${temp}_aroma_${fun}
-			3dTstat -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-aroma_${fun}+tlrc ${PrepDir}/${sub}/afni/${sub}_${temp}_aroma_${fun}+tlrc
-			3dcalc -a ${PrepDir}/${sub}/afni/${sub}_${temp}_aroma_${fun}+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-aroma_${fun}+tlrc \
+				# Clean up mean image
+				rm -f ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-preproc_${fun}+tlrc*
+				#
+				# Now copy over the brain mask into brik/head format
+				3dcopy ${PrepDir}/${sub}/func/${sub}_task-${fun}_space-${temp}_desc-brain_mask.nii.gz ${PrepDir}/${sub}/afni/${sub}_${temp}_brainmask_${fun}
+				#
+				# Finally copy over the AROMAe'd data and scale like before
+				3dcopy ${PrepDir}/${sub}/func/${sub}_task-${fun}_space-${temp}_desc-smoothAROMAnonaggr_bold.nii.gz ${PrepDir}/${sub}/afni/${sub}_${temp}_aroma_${fun}
+				3dTstat -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-aroma_${fun}+tlrc ${PrepDir}/${sub}/afni/${sub}_${temp}_aroma_${fun}+tlrc
+				3dcalc -a ${PrepDir}/${sub}/afni/${sub}_${temp}_aroma_${fun}+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-aroma_${fun}+tlrc \
 				-expr 'min(200, a/b*100)*step(a)*step(b)' \
 				-prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-aroma_${fun}+tlrc
-			rm -f ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-aroma_${fun}+tlrc*
-			#
-			#
-			#
-			# Now mask the scaled images. First have to rescale mask
-			# First, the aroma, which has different dimensions, so needs to be rescaled
-			3dresample -master ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-aroma_${fun}+tlrc -prefix ${PrepDir}/${sub}/afni/TempMask -input ${PrepDir}/${sub}/afni/${sub}_${temp}_brainmask_${fun}+tlrc
-			3dcalc -a ${PrepDir}/${sub}/afni/TempMask+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-aroma_${fun}+tlrc -expr 'a*b' -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_masked-scaled-aroma_${fun}
-			rm -f ${PrepDir}/${sub}/afni/TempMask+tlrc*
-			#
-			# Then the preproc, which doesn't need scaling
-			3dcalc -a ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-preproc_${fun}+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_brainmask_${fun}+tlrc -expr 'a*b' -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_masked-scaled-preproc_${fun}
-			#
+				rm -f ${PrepDir}/${sub}/afni/${sub}_${temp}_mean-aroma_${fun}+tlrc*
+				#
+				#
+				#
+				# Now mask the scaled images. First have to rescale mask
+				# First, the aroma, which has different dimensions, so needs to be rescaled
+				3dresample -master ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-aroma_${fun}+tlrc -prefix ${PrepDir}/${sub}/afni/TempMask -input ${PrepDir}/${sub}/afni/${sub}_${temp}_brainmask_${fun}+tlrc
+				3dcalc -a ${PrepDir}/${sub}/afni/TempMask+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-aroma_${fun}+tlrc -expr 'a*b' -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_masked-scaled-aroma_${fun}
+				rm -f ${PrepDir}/${sub}/afni/TempMask+tlrc*
+				#
+				# Then the preproc, which doesn't need scaling
+				3dcalc -a ${PrepDir}/${sub}/afni/${sub}_${temp}_scaled-preproc_${fun}+tlrc -b ${PrepDir}/${sub}/afni/${sub}_${temp}_brainmask_${fun}+tlrc -expr 'a*b' -prefix ${PrepDir}/${sub}/afni/${sub}_${temp}_masked-scaled-preproc_${fun}
+				#
+			fi
 			# Now convert the confound regressors file to afni compatible
 			if [ -f ${PrepDir}/${sub}/afni/${sub}-${fun}-MotionSummary.csv ]; then
 				echo "Looks like motion extraction has already occurred"
